+++ CHANGELOG.md
@@ -2,13 +2,27 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.5.2-SNAPSHOT] - 16.09.2025.
+## [0.5.3-SNAPSHOT] - 18.09.2025.
 
 ### Added
 
 - Logic for static serving /.well-known/dspace-version document
 - New properties for DSpace version document; stored in database; logic for retrieving and parsing properties
 - New set of GitHub Action that will be used to test Connector related features
+
+## [0.5.2-SNAPSHOT] - 17.09.2025.
+
+### Added
+
+- Added HTTP-Push transfer format
+- Added HTTP-Push transfer diagram
+
+### Changed
+
+- Since the consumer won't know when the data will be uploaded, the COMPLETED state will be used as the indicator.
+  After successfully finishing the upload, the provider will send a TransferProcessCompleted message.
+  The isDownloaded flag will remain in the code for the moment but might be removed in the future.
+- Changed HTTP-Pull transfer diagram
 
 ## [0.5.1-SNAPSHOT] - 08.08.2025.